Package project with CLI tools and vision bridge
